fix: consolidate files, avoid import cycle
